HTML Timetable
A light blue background color is set for the <body> element using the bgcolor attribute.

The time table is displayed within a table (<table>) element.

The table has a border width of 3 pixels and a cellspacing of 4 pixels.
The table is aligned to the center using the align attribute.
The table heading is displayed in a centered heading level 1 (<h1>) element with the text "COLLEGE TIME TABLE". The heading is colored cyan using the inline style attribute style="color:cyan".
The table rows (<tr>) and table header cells (<th>) define the days of the week and time slots.

The time slots are specified in each table header cell.
The table header cells for the days of the week are displayed in bold.
The table data cells (<td>) represent the subjects and activities for each day and time slot.

The subjects are displayed in various colors using inline styles (style="color:blue", style="color:yellow", style="color:red", etc.).
The rowspan attribute is used in the lunchtime slot to span across multiple rows.
The subjects and activities in the table represent a sample college time table for different days of the week.